fix: refactor spawnsize
