+++ src/helpers/files.ts
@@ -7,7 +7,7 @@ import { UploaderArgs } from '../types'
 import { logError, UploadLogger } from './logger'
 import { runExternalProgram } from './util'
 import micromatch from "../vendor/micromatch/index.js";
-import { SPAWNPROCESSBUFFERSIZE } from './constansts'
+import { SPAWNPROCESSBUFFERSIZE } from './constants'
 
 export const MARKER_NETWORK_END = '\n<<<<<< network\n'
 export const MARKER_FILE_END = '<<<<<< EOF\n'
