+++ test/providers/provider_bitrise.test.ts
@@ -1,10 +1,10 @@
 import td from 'testdouble'
 import childProcess from 'child_process'
-import { SPAWNPROCESSBUFFERSIZE } from '../../src/helpers/util'
+import { SPAWNPROCESSBUFFERSIZE } from '../../src/helpers/constants'
 import { IServiceParams, UploaderInputs } from '../../src/types'
 import { createEmptyArgs } from '../test_helpers'
 
-import * as providerBitrise from '../../src/ci_providers//provider_bitrise'
+import * as providerBitrise from '../../src/ci_providers/provider_bitrise'
 
 describe('Bitrise Params', () => {
   afterEach(() => {
